Rename LinkTokenPage to TokenListPage
