+++ client/src/pages/TokenListPage/components/InfoMessage.tsx
